refactor(webhookpub): extract shared fan-out core into fanout_core.go
Pure intra-package move (no behavior change): the fan-out logic shared by
the two engines that drain the webhook_events outbox — fanOutEventCore,
insertPendingBatchTx, generateDeliveryID, derefString, poolBeginFunc, the
leasedEvent row-shape, and the DeliveryEnqueuer + identityReader
interfaces — moves out of worker.go into a new fanout_core.go.

worker.go is now ENTIRELY the legacy LISTEN/NOTIFY OutboxWorker (326 lines,
was ~500), so the post-cutover deletion becomes "rm worker.go" instead of
carving the shared half out of a mixed file. fanout_core.go (223 lines) is
the durable surface both the legacy OutboxWorker and the River FanOutWorker
depend on, in one place.

Compiler-verified move; all webhookpub tests pass unchanged.
